+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.14.3] - 2020-01-10
+
+* JPMS bugfix
+
+## [0.14.2] - 2020-01-10
+
+* Opens package to configure the Exception Handler by reflection
+
 ## [0.14.1] - 2020-01-09
 
 * Add new properties on Quarkus extension: `enabled`